backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Pydantic / Settings ───────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Auth ──────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── LLM / OpenAI ─────────────────────────────────────────────
openai==1.30.1

# ── RAG / Embeddings ─────────────────────────────────────────
sentence-transformers==3.0.0
rank-bm25==0.2.2

# ── Vector DB — Qdrant ───────────────────────────────────────
qdrant-client==1.9.1

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.30

# ── Data Generation ───────────────────────────────────────────
faker==25.0.0

# ── Observability ─────────────────────────────────────────────
langfuse==2.30.0

# ── Utilities ─────────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
tiktoken==0.7.0

# ── Testing ───────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
